4.21: laajennus step 9

+++ assignments/osa4/blogilista/controllers/blogs.js
@@ -48,8 +48,28 @@ blogsRouter.post('/', async (request, response, next) => {
 //4.13
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    //step 9
+    const decodedToken = jwt.verify(request.token, config.SECRET)
+
+    if (!decodedToken.id)
+    {
+      return response.status(401).json({error: 'token invalid'})
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog){
+      return response.status(404).json({error:'blog not found'})
+    }
+
+    if (blog.user.toString() === decodedToken.id.toString()){
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+    } else {
+      return response.status(403).json({error:'permission denied'})
+    }
+    //await Blog.findByIdAndDelete(request.params.id)
+    //response.status(204).end()
   } catch (exception) {
     next(exception)
   }
